requirements-dev.txt
# Development dependencies
-r requirements.txt

# Crypto support (required for tests)
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
requests-mock>=1.11.0

# Code quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
types-requests>=2.31.0

# Build
build>=1.0.0
twine>=4.0.0
